+++ src/migrations/Install.php
@@ -44,6 +44,18 @@ class Install extends Migration
 				'uid' => $this->uid()
 			]);
 		}
+
+		if (!$this->db->tableExists('{{%linkvault_reports}}')) {
+			$this->createTable('{{%linkvault_reports}}', [
+				'id' => $this->primaryKey(),
+				'criteria' => $this->text(),
+				'orderBy' => $this->string(50),
+				'sort' => $this->string(4),
+				'dateCreated' => $this->dateTime()->notNull(),
+				'dateUpdated' => $this->dateTime()->notNull(),
+				'uid' => $this->uid()
+			]);
+		}
 	}
 
 	private function _upgradeFromCraft2()
@@ -98,6 +110,9 @@ class Install extends Migration
 		}
 		if ( $this->db->tableExists('{{%linkvault_customfields}}') ) {
 			$this->dropTable('{{%linkvault_customfields}}');
+		}
+		if ( $this->db->tableExists('{{%linkvault_reports}}') ) {
+			$this->dropTable('{{%linkvault_reports}}');
 		}
 	}
 }